+++ bostonBusMap/src/boston/Bus/Map/math/Geometry.java
@@ -51,7 +51,7 @@ public class Geometry {
 	 * @param lon2 longitude in radians
 	 * @return distance in miles
 	 */
-	public static float computeCompareDistance(float lat1,	float lon1, float lat2, float lon2)
+	public static float computeCompareDistance(double lat1, double lon1, double lat2, double lon2)
 	{
 		double dist = (((1.0 - Math.cos(lat1 - lat2)) * 0.5) + Math.cos(lat1) * Math.cos(lat2) * 
 				((1.0 - Math.cos(lon1 - lon2)) * 0.5));
