+++ bostonBusMap/src/boston/Bus/Map/util/Constants.java
@@ -3,6 +3,6 @@ package boston.Bus.Map.util;
 public class Constants {
 
 	public static final int E6 = 1000000; 
-	public static final float InvE6 = 1/(float)E6;
+	public static final double InvE6 = 1/(double)E6;
 	
 }
